Make 64bit conformable.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@664 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ vTextField.c
@@ -908,7 +908,7 @@ RequestSelection(Widget aw, XtPointer client, Atom * selection, Atom * type,
 
     ClearHighlight(w);
     savex = w->text.OldCursorX;
-    w->text.CursorPos = (int) client;
+    w->text.CursorPos = (long) client;
 #ifdef DEBUG_TF
     printf("RequestSelection: inserting %s length=%d at pos: %d\n",
       (char *) value, (int) (*length), w->text.CursorPos);
